src/requirements.txt
langchain==0.3.7
langchain-core==0.3.13
langgraph==0.1.15

openai==1.52.2
langchain-openai==0.2.6
azure-identity==1.17.1
azure-ai-ml==1.21.0

python-dotenv==1.0.1

pydantic==1.10.15
regex==2024.9.11

rich==13.9.2